requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
gunicorn==19.7.1
idna==2.6
lxml==4.2.1
python-telegram-bot==10.1.0
requests==2.18.4
urllib3==1.22
